Updated the app function and added a _README
The index.js file has had the  old logic for copying and moving files
between the template and app directory rewritten from generator-
playbook. Less code and easier to maintain.

Added a README.md to the template directory that will be copied over
to the app directory so users can get started using jekyllized.

Updated the changelog to reflect this.