Removed most of the logic for choosing a template
This is to start working towards having a sane and good default
template to use instead of multiple. See #14 for more information about
this. This update removed most of the conditionals associated with
H5BP and its files and such.

Removed the configurations for SASS/Compass from the Gruntfile.js to
get even closer to completing having a lone default theme. See #14 again.

Removed Compass and added CSSCSS from the Gruntfile configurations

Removed and added several different configs

Removed all code for Compass and added the code for CSSCSS, did some
other minor cleaning as well. See #14 and #15 for more information.

Changed some leftover names (jsPre) that weren't changed earlier

This fixed #7.

Removed the conditionals for H5BP, see #14

Updated support for Coffeescript and added an app.file to the default
location for the Cofffee scripts.