More leftover old parameter names in files
For this commit projectName was a leftover that should've been
projectname (note the lowercase). Fixed with this commit. Fixes #7.

Further renaming of old parameters and such

Leftover of jekPyg that should've been jekyllPygments. This is a
continuation of issue #7.

Renamed author into owner.name and furter edits

Further renamings of parameters