Restructuring of index.js to fix path.js errors
With the new logic everytime the gems are finished bundling the
generator throws and error about path.js. This is apparently a
bug with bower, but the fixes I've tried so far doesn't work.

See #1, but this will need more fixing.

Renamed template-default to template

As we're removing all other templates we will only need one template,
and having it named default is a bit redundant. See #14.

grunt-autoprefixer got accidentally renamed

This fixes that error. See #7 for more.