Added a very basic theme and fixed running tests
There's now a very, very basic theme implemented for the generator,
however it still leaves a lot to be done. At least there's something
there now. It uses H5BP as a base but needs further updates, especially
for the logic Jekyll requires. This fixes #25.

It also fixes being able to run npm test on the repo, so now Travis
will be able to test it and you can test it locally. This fixes #24 and
fixes #4.