Added .npmignore, welcome message, updated deps
Added a .npmignore file to not push everything to NPM when I run
publish on this generator.

The license in the package.json file now links to the license in the
repo. This fixes #28. Also updated the version required for chalk,
mocha, globule, shelljs and yeoman-generator.

Lastly I made a custom Yeoman welcome message instead of the default
on that pops up and added some nice colors. This fixes #27.