Further updates on the restructuring of index.js
This is a continuation of the major overhaul of index.js and
Gruntfile.js, see #30 for more information. This just adds more
parameters and variables that are currently empty but will be used
later for generating the site. Added most of the empty logic for #32,

Also fixed running tests on the site. And updates travis.yml according
to their docs. Updated CHANGELOG to reflect this.