Major restructuring and update of index.js
Made some major rewrites on how all the prompts looks and works as well
as fixed it so that when choosing what kind of preprocessor you want it
actually shows the correct default directory to place it in instead of
having it be empty. Did some other minor fixes, moved the welcome into
it's own generator, fixed running tests on it while using split and
toLowerCase on variables.

This updates #34, #32, #33 and #31 and for a complete overview see #30