Continuation of adding Gulp and updating index.js
Removed the Bitdeli and Waffle.io badge from the Readme and updated the
index.js to follow the changes needed for this refactoring. See #30 for
more. Renamed the config.json file for some clarity.

Removed the grunt packages and added the gulp packages, added some info
about the asset structure in the gulpfile and minor fixes for the
config.json file.

Also did some fixing on the error reporting for the app. It should now
not install all the changes and instead just run them.