More changes to help go with the Gulp overhaul
Updates in the atom feed and sitemap for layout changes in Jekyll 2.0
and renaming off pygments in the configuration. Also did some renames of
the stylesheets for some minor clarity and personal preference. And
obviously even more changes to the gulpfile, it's now a lot cleaner than
before and has more functionality. Huzzah.