Renaming tasks and more fixes for the gulpfile
Renamed a bunch of tasks to make it more clear what they are doing/are
used for. For example jekyll:serve is now jekyll:dev to show that it's
used when developing. Also added BrowserSync notifications for when
Jekyll is rebuilt and CSS is injected.

Removed a few tasks, the copy and csslint task is now removed and a new
copy task is added for copying images, media assets and .txt/.xml files
to the final output folder. Lots of minor fixes for issue #48.
gulp-useref no longer doesn't copy the CSS file to the 'site' folder nor
should cleaning the directories throw an error anymore.