Major overhaul of the gulpfile and tests plus plus
With this commit there are lots of crazy things happening, the index.js
file has been more or less rewritten and the tests has been fixed ever
so slightly to actually do something. Also changed the names for some
variables here and there and other minor changes.

However, this also breaks because apparently there's an paranthesis
that makes the generator throw an error that comes from the gulpfile.
Weird. #48 will have to be updated because this really breaks things.