Getting there now
The Gulpfile now can upload and gzip files for Amazon S3 and can update
the root object in Cloudfront distributions. Added the needed packages
to the package.json file, added a distribution ID needed for Cloudfront
to aws-credentials.json and made changes accordingly to the tests. See
issue #51 for more.