Updated packages and added more tests
Added more tests to help with developing stuff that doesn't break and
also rewrote them in such a way that they won't output all the various
messages for the prompts to make reading the test output a lot easier.
Added a test that checks that all the gulp tasks are there, a test that
check that the aws-credentials.json file is created and likewise for
Rsync.

Also updated the packages in the package.json file.