Updated tests, began work on GH Pages uploading
Updated some names in the files, owner -> author and updated the
directories that tests are created it. Also updated the tests to not
include all the prompts that they don't need. Added a test that checks
that Jekylls options are correctly set and discovered a bug with
selection pagination that was fixed.

Also added a small beginning for support of Github Pages and added
extensions to Redcarpet for better output of HTML and support for
excerpts.

Updated the NPM packages too to be more up to date and updated the
index.js to be more up to par.