Added support for GitHub Pages
GitHub pages are now supported, but for the moment only personal
repositories, this is something that is open for change for the moment.
This closes #61. Also removed the gh-branch on the official repository
as it was unused (for the moment), so this closes #56.

Also updated the tests so they don't mock everything but only the things
that are required, this is mostly a change for convenience. With this
testing should be pretty okay. This closes #66 and closes #60.