Updated documentation, directories and tests
Updated the README to reflect the most recent changes and reword the
introduction, updated the CHANGELOG with all the updates.

Moved the output for Jekyllized into a dist folder instead of a site
folder to be more in line with other tools and updated the test for gulp
tasks to include the new tasks.