Archives, improved generation and more
Updated the _config files to help with incremental regeneration so you
don't have to fully rebuild the site all the time. Also added the
jekyll-archives gem to Jekyll with proper settings and some small
beginnings for the layouts.

Updated the atom.xml file to be more up to par and validate correctly
and added a small "Hello world!" log to one of the JS files so you can
see that it is included properly in the JS console.