Updated all the tests and fixed the deploy task
Added a bunch more tests that checks for the gulpfile tasks, that the
correct packages are installed in the package.json file and removed the
test-util.js file as it's no longer needed.

Also fixed the gulpfile always containing a deploy task even when you
selected that you didn't use it and added tests that checks for it.