Updated tests, packages and Travis settings
Moved to JSCS for checking the JS style to be more consistent and find
dumb errors before they are committed. Means that there a bunch of
aesthetic changes to them. Also updated the tests to check for the
functions that were added as we migrate to Gulp 4.

Updated the packages and moved them from peerDeps to devDeps as NPM will
apparently stop using peerDeps in NPM 3.0.

Updated JShint settings and tests: We're now testing for both node 0.10, 0.12
and iojs as well. However we moved from the default Travis NPM version to the
latest NPM version and fixed some minor errors with Travis and how tests are
run.