More updates to the gulp subgenerator and tests
Made the subgenerator properly scaffold a gulpfile and packages now and
updated the tests accordingly for the gulpfile and moved most of the
small tests into a boilerplate test file. Also updated the gulpfile at
the root directory to look for JS errors in the generated gulpfiles and
run a singular test task so even when there are errors it will actually
fail. Neat.

Also updated yeoman-generator to a newer version (0.20.1) for some
templating goodness.