Beginning of Bower
This commit adds support for Bower in a very rudimentary state, it
currently does mostly all the things you want it to do but I'd like for
it to copy the production versions for the final build step, will have
to look into that. Otherwise it works. Also updated the tests to verify
that packages are installed correctly and such.

Fixes #79.