Overhauled the Gulpfile
Fixed a bunch of minor problems with generating JS and CSS, removed the
cache buster for now because of how much of a pain in the ass it has
been, renamed some JS files and removed Useref and now use gulp-inject
instead do directly inject assets. Also updated tests accordingly.