Update code style, continue overhaul of gulpfile
Updated to use the XO style guide (with spaces obviously) and fixed all
the errors I could figure out and fix with it.
Removed a lot of packages that weren't being used at all and updated
their versions and updated the tests accordingly. Also tweaked the tests
for packages and gulp tasks to be easier to add tasks etc.
Continued my overhaul of the gulpfile, added new tasks while developing
and updated the production tasks. Still a major WIP.