Fixes sondr3/generator-jekyllized#127
This update to the gulpfile and Jekyll configuration streamlines
the build process, described below:

1. Check for errors/warnings in JavaScript and Jekyll configuration
2. Clean the whole project (.tmp and dist directories)
3. Build the assets (images, CSS, JS, fonts) in .tmp/assets-built
4. Copy files to be processed by Jekyll to .tmp/jekyll-preprocessed-src
5. Inject head and footer to .tmp/jekyll-preprocessed-src
6. Run Jekyll against ./tmp/jekyll-preprocessed-src to .tmp/jekyll-built
7. Copy built assets to dist
8. Copy built Jekyll site to dist
9. Run html task if the --prod switch is activated

While serving the site (from .tmp/assets-built and .tmp/jekyll-built),
gulp watches for any change in assets (including favicon), Jekyll contents
(md, txt, xml, html) and propagates the changes to the relevant .tmp
directory. In case of images and favicon, a specific clean operation is
performed to handle deletions.

Every change is also propagated to the dist directory so it always
represent the final state of the build. This is valid for both the default
and the build tasks.

The rebuild task has been removed as it's not relevant anymore.