Copy to a temporary directory, ES2015 updates
This is so that your git commit history isn't disturbed by tons of
changes to your header and footer file, now it's all done in the
temporary directory first.

Also changed the file to use ES2015 syntax without having to rely on
Babel, this speeds up the gulpfile quite a bit.

This closes #131 and fixes #127, fixes #128 and fixes #129.