Move back to a single gulpfile
We're going back to a single gulp file instead of using several ones,
although it worked decently well I wasn't all too happy with how it
worked. So we're back to how we did it with the README. Not sure if it's
the best way to do this but I think it'll be fine.