Added a very barebones Jekyll subgenerator
Included the same useful gems as we did in the previous version as well as the CommonMark extension, set the source and destination paths but otherwise it's a clean and default Jekyll installation.

Also fixed how the installation is run with the runInstall() command instead of hacking something together on my own, which coincidentally also gives us 100% coverage still!